kaggle/src/utils/logger.py
from dotenv import load_dotenv
import os
import logging
from pathlib import Path

load_dotenv()
LOGS_DIR = os.getenv('LOGS_DIR', 'logs')

def setup_logging(log_dir: str = LOGS_DIR) -> logging.Logger:
    log_dir = Path(log_dir)
    log_dir.mkdir(parents=True, exist_ok=True)
    
    log_file = log_dir / "latest.log"
    
    logger = logging.getLogger("TextClassification")
    logger.setLevel(logging.INFO)
    
    if not logger.handlers:
        formatter = logging.Formatter('%(asctime)s - %(name)s - %(levelname)s - %(message)s')
        
        file_handler = logging.FileHandler(log_file)
        file_handler.setLevel(logging.INFO)
        file_handler.setFormatter(formatter)
        logger.addHandler(file_handler)
        
        console_handler = logging.StreamHandler()
        console_handler.setLevel(logging.INFO)
        console_handler.setFormatter(formatter)
        logger.addHandler(console_handler)
    
    return logger
