kaggle/requirements.txt
numpy
pandas
scikit-learn
matplotlib
seaborn
python-dotenv
